Added break point for OutOfRetries error handler
In the event the Fitbit server could not be contacted in the middle of the script it would fail to POST and throw a requests library error for out of retries.  Wrapped it in a try/except block thanks to some code help from Anecdata.  It has helped.  Currently the script doesn't fail but it can get stuck if the mobile app doesn't update from the Fitbit device via BLE.  The display will simply continue to show the same last 15 logged values.  The same last values were confirmed in Fitbits online dashboard.  It's not a script error it's non-communication between Fitbit device and mobile app.  Aha! Took a while to find that culprit.